Clean up or sending request
